introduce interfaces for clients
