+++ boudicca.base/enricher-client/src/main/kotlin/base/boudicca/api/enricher/EnricherClient.kt
@@ -1,57 +1,7 @@
 package base.boudicca.api.enricher
 
-import base.boudicca.enricher.openapi.api.EnricherApi
-import base.boudicca.enricher.openapi.model.EnrichRequestDTO
 import base.boudicca.model.Event
-import base.boudicca.openapi.ApiClient
-import base.boudicca.openapi.ApiException
-import io.opentelemetry.api.GlobalOpenTelemetry
-import io.opentelemetry.api.OpenTelemetry
-import io.opentelemetry.instrumentation.javahttpclient.JavaHttpClientTelemetry
-import java.net.http.HttpClient
-import base.boudicca.enricher.openapi.model.Event as EnricherOpenApiEvent
 
-class EnricherClient(
-    private val enricherUrl: String,
-    private val otel: OpenTelemetry = GlobalOpenTelemetry.get(),
-) {
-    private val enricherApi: EnricherApi
-
-    init {
-        check(enricherUrl.isNotBlank()) { "you need to pass an enricherUrl!" }
-        val apiClient =
-            object : ApiClient() {
-                override fun getHttpClient(): HttpClient? =
-                    JavaHttpClientTelemetry
-                        .builder(otel)
-                        .build()
-                        .wrap(super.getHttpClient())
-            }
-        apiClient.updateBaseUri(enricherUrl)
-
-        enricherApi = EnricherApi(apiClient)
-    }
-
-    fun enrichEvents(events: List<Event>): List<Event> {
-        try {
-            return enricherApi
-                .enrich(EnrichRequestDTO().events(events.map { mapToEnricherEvent(it) }))
-                .map { toEvent(it) }
-        } catch (e: ApiException) {
-            throw EnricherException("could not reach enricher: $enricherUrl", e)
-        }
-    }
-
-    private fun toEvent(enricherEvent: EnricherOpenApiEvent): Event = Event(enricherEvent.name!!, enricherEvent.startDate!!, enricherEvent.data ?: mapOf())
-
-    private fun mapToEnricherEvent(event: Event): EnricherOpenApiEvent =
-        EnricherOpenApiEvent()
-            .name(event.name)
-            .startDate(event.startDate)
-            .data(event.data)
+interface EnricherClient {
+    fun enrichEvents(events: List<Event>): List<Event>
 }
-
-class EnricherException(
-    msg: String,
-    e: ApiException,
-) : RuntimeException(msg, e)
