+++ boudicca.base/eventcollector-client/src/main/kotlin/base/boudicca/api/eventcollector/runner/RunnerIngestionInterface.kt
@@ -1,7 +1,7 @@
 package base.boudicca.api.eventcollector.runner
 
 import base.boudicca.api.eventcollector.Configuration
-import base.boudicca.api.eventdb.ingest.EventDbIngestClient
+import base.boudicca.api.eventdb.ingest.DefaultEventDbIngestClient
 import base.boudicca.model.Event
 import io.opentelemetry.api.OpenTelemetry
 
@@ -17,7 +17,7 @@ fun interface RunnerIngestionInterface {
                     ?: error { "you need to specify the boudicca.ingest.auth property!" }
             val user = userAndPassword.split(":")[0]
             val password = userAndPassword.split(":")[1]
-            val eventDb = EventDbIngestClient(eventDbUrl, user, password, otel)
+            val eventDb = DefaultEventDbIngestClient(eventDbUrl, user, password, otel)
             return BoudiccaRunnerIngestionInterface(eventDb)
         }
     }
